Commit 8
Most updated version of the streamlit application. The only changes were made at the bottom of the code with the buttons at the bottom, but it says the whole thing is different because I recopied the code from Visual Studio Code over to Jupyter notebook so it thought I changed each block of code.